requirements.txt
opencv-python-headless
websockets
numpy
python-dotenv